Replace datatype from int to Integer etc. to support PUT with multiple given params only.

+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/dao/RoleDao.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface RoleDao {
 
-    List<Role> queryRolesByUserId(int userId);
+    List<Role> queryRolesByUserId(Long userId);
 
 }
